removed unnecessary type cast

+++ src/player/interface.rs
@@ -37,7 +37,7 @@ impl MusicPlayer {
             .iter()
             .position(|r| r.clone().uniqueid == cache.song_id)
             .unwrap_or(0);
-        let current_object = list[count as usize].clone();
+        let current_object = list[count].clone();
         // list should inherite from cache at some point. not worried now tho
         MusicPlayer {
             list,
